Update directory structure and V0.0.2 of Mote Proj
src at TLD was renamed to projects to reflect the organisation
of the repository - and ultimately remove confusion (as the src
directory is repeated in every project directory).

Gitignore was updated to reflect Atmel Studio 7 files.

V0.0.2 of C7C.Atamo.Dusty.Mote was added; this integrates the
Arduino build files with Atmel Studio 7