Fixes #6
The Dusty default Flash image was copied, and modified using
Eterna's fuse utility. Of key note, was the Baud rate (from 115.2k
to 9600). This is because of a low frequency ripple that is created
at high baud rates.

We also introduce a compareFlash utility (Python) to compare flash
images.

Note some issues remain, in particular the NETID in the DP does
not actually affect the joining process, despite the fact that
it should