Sample unit tests - LOCAL BUILD ONLY
Software unit tests for sample* functions implemented and working in a
test environment - NOT BUILDABLE FOR duinoPRO. Specificly, unit tests for:
 - sample(uint8_t)
 - sample_time()
 - sample_sensor()
 - sample_diagnostic()

A file "sample_offline_tests.cpp" has been added for running the above
unit tests. The command for compiling this is contained at line five of
the file. Simply build and run this to execute each unit test. Results
are printed to the console.

Unit tests are designed to be run in an *offline* environment, not the
duinoPRO. As such, the following header files have been commented out:
 - duinoPRO.h
 - AccelMagno.h
 - ITimo.h
 - Arduino.h
Further, the following functions have been modified to give a
(deterministic) output for a local build for testing purposes:
 - sample_diagnostic: constant Vbat output
 - sensor_read: constant sensor readings
 - dn_time_ms: constant time output
Other notes on unit tests:
 - Some other modifications were required to facilitate local building.
   All instances are noted by a comment containing the word "TESTING" in
   all capitals.
 - printf calls are used extensively in the unit test code for
   convenience. These cannot be executed on the duinoPRO.

Miscellaneous:
 - Corrected maximum payload length in frame.h
 - Corrected sensor type code from 0x0 to 0x00 (1 byte)
 - Notation updated in TestDpFraming.cpp
 - Bug fix in sensor_read (incorrect computation of buffer size)
 - In process of unit testing, corrected bugs affecting order in which
   data/diagnostic bytes are saved to the payload in sample_* functions.